Actualizados tests

+++ src/main/java/Aritmetica.java
@@ -13,7 +13,7 @@ public class Aritmetica {
   }
 
   public static float division (int dividendo, int divisor) {
-        return (dividendo/(float)divisor);
+        return (float)(dividendo/(float)divisor);
   }
 
 }
